shared/scriptlet: Move scriptlet utils to shared
