+++ shared/scriptlet/marshal_test.go
@@ -1,4 +1,4 @@
-package marshal
+package scriptlet
 
 import (
 	"fmt"
